requirements.txt
asn1crypto==1.5.1
boto3==1.40.38
botocore==1.40.38
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==46.0.0
et_xmlfile==2.0.0
filelock==3.19.1
greenlet==3.2.4
idna==3.10
jmespath==1.0.1
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
platformdirs==4.4.0
pyarrow==21.0.0
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
snowflake-connector-python==3.17.4
sortedcontainers==2.4.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
